moving sdks/fabric/go-cli to samples/fabric/clis/go/fabric-cli
