+++ samples/fabric/clis/go/fabric-cli/Makefile
@@ -1,0 +1,9 @@
+all: clean build
+
+.PHONY: clean
+clean:
+	rm -rf ./bin
+
+.PHONY: build
+build:
+	go build -o ./bin/fabric-cli fabric-cli.go
