+++ samples/fabric/clis/go/fabric-cli/config.template.json
@@ -1,0 +1,16 @@
+{
+  "network1": {
+    "connProfilePath": "<PATH-TO-WEAVER>/tests/network-setups/fabric/shared/network1/peerOrganizations/org1.network1.com/connection-org1.yaml",
+    "relayEndpoint": "localhost:9080",
+    "mspId": "Org1MSP",
+    "channelName": "mychannel",
+    "chaincode": "simpleasset"
+  },
+  "network2": {
+    "connProfilePath": "<PATH-TO-WEAVER>/tests/network-setups/fabric/shared/network2/peerOrganizations/org1.network2.com/connection-org1.yaml",
+    "relayEndpoint": "localhost:9083",
+    "mspId": "Org1MSP",
+    "channelName": "mychannel",
+    "chaincode": "simpleasset"
+  }
+}
