+++ samples/fabric/clis/go/fabric-cli/data/credentials/network1/verification-policy.json
@@ -1,0 +1,1 @@
+{"securityDomain":"network1","identifiers":[{"pattern":"mychannel:simplestate:Read:*","policy":{"type":"Signature","criteria":["Org1MSP"]}}]}
