+++ samples/fabric/clis/go/fabric-cli/data/starSize.json
@@ -1,0 +1,27 @@
+{
+  "Arcturus": "17.671",
+  "Betelgeuse": "617.1",
+  "Canopus": "49.395",
+  "Deneb": "141.23",
+  "Electra": "4.2159",
+  "Furud": "2.713",
+  "Giausar": "27.065",
+  "Hadar": "5.983",
+  "IZAR": "18.412",
+  "Jabbah": "6.033",
+  "Kuma": "1.386",
+  "Lesath": "4.244",
+  "Maia": "4.202",
+  "Naos": "9.74",
+  "Omicron1 Eridani": "2.5741",
+  "Peacock": "3.3602",
+  "Rotanev": "2.752",
+  "Sadalsuud": "34.8",
+  "Talitha": "1.288",
+  "UNUKALHAI": "8.3484",
+  "VEGA": "1.6432",
+  "Wasat": "1.649",
+  "Xamidimura": "4.774",
+  "Yildun": "1.948",
+  "Zosma": "1.4888"
+}
