+++ samples/fabric/clis/go/fabric-cli/data/stars.json
@@ -1,0 +1,28 @@
+{
+  "a": "Arcturus",
+  "b": "Betelgeuse",
+  "c": "Canopus",
+  "d": "Deneb",
+  "e": "Electra",
+  "f": "Furud",
+  "g": "Giausar",
+  "h": "Hadar",
+  "i": "IZAR",
+  "j": "Jabbah",
+  "k": "Kuma",
+  "l": "Lesath",
+  "m": "Maia",
+  "n": "Naos",
+  "o": "Omicron1 Eridani",
+  "p": "Peacock",
+  "q": "No star with letter q",
+  "r": "Rotanev",
+  "s": "Sadalsuud",
+  "t": "Talitha",
+  "u": "UNUKALHAI",
+  "v": "VEGA",
+  "w": "Wasat",
+  "x": "Xamidimura",
+  "y": "Yildun",
+  "z": "Zosma"
+}
