+++ samples/fabric/clis/go/fabric-cli/helpers/.gitignore
@@ -1,0 +1,1 @@
+wallet
