+++ samples/fabric/clis/go/fabric-cli/helpers/testdata/example/peerOrganizations/org1.example.com/users/User1@org1.example.com/msp/keystore/priv_k
@@ -1,0 +1,6 @@
+-----BEGIN PRIVATE KEY-----
+MIGTAgEAMBMGByqGSM49AgEGCCqGSM49AwEHBHkwdwIBAQQggkuKP0YNrbuilpFf
+0F/I+3At9LZh6EysU8lVBuy+cregCgYIKoZIzj0DAQehRANCAAQ3NMOS6YpCyFKJ
+jgKYCP6eQYUG91jdhoQK+8Ufhy0/V/CVdJj/Exe89yzAqKfLzb9tc6MuWOYLwPRD
+sF3d8qsw
+-----END PRIVATE KEY-----
