--- sdks/fabric/go-cli/.gitignore
@@ -1,3 +1,0 @@
-wallet
-go-sdk
-keystore
