--- sdks/fabric/go-cli/data/credentials/network1/access-control.json
@@ -1,11 +1,0 @@
-{
-  "securityDomain": "network1",
-  "rules": [
-    {
-      "principal": "Org1MSP",
-      "principalType": "ca",
-      "resource": "mychannel:simpleasset:Read:*",
-      "read": true
-    }
-  ]
-}
