--- sdks/fabric/go-cli/data/credentials/network1/verification-policy.json
@@ -1,1 +1,0 @@
-{"securityDomain":"network1","identifiers":[{"pattern":"mychannel:simplestate:Read:*","policy":{"type":"Signature","criteria":["Org1MSP"]}}]}
