--- sdks/fabric/go-cli/data/interop/accessControlTemplate.json
@@ -1,11 +1,0 @@
-{
-  "securityDomain": "",
-  "rules": [
-    {
-      "principal": "<mspid>",
-      "principalType": "ca",
-      "resource": "mychannel:simplestate:Read:*",
-      "read": true
-    }
-  ]
-}
