--- sdks/fabric/go-cli/data/interop/verificationPolicyTemplate.json
@@ -1,11 +1,0 @@
-{
-    "securityDomain": "<network-id>",
-    "identifiers": [
-      {
-        "pattern": "mychannel:simplestate:Read:*",
-        "policy": {
-          "type": "Signature",
-          "criteria": []
-        }
-      }]
-  }
