--- sdks/fabric/go-cli/helpers/.gitignore
@@ -1,1 +1,0 @@
-wallet
